Delegate output to snyk-iac-test
If the environment variable SNYK_IAC_TEST_OUTPUT is set to 1,
snyk-iac-test will inherit the stdin and stdout of the Snyk CLI. This
allows snyk-iac-test to take care of the spinner and produce its own
output. Whatever snyk-iac-test prints to stdout will be immediately
visible to the user of the Snyk CLI.

The stderr of snyk-iac-test is not inherited from Snyk, and is instead
buffered. The stderr from snyk-iac-test is only visible to the user if
they provide the -d flag on the command line.

The SNYK_IAC_TEST_OUTPUT environment variable will be inherited by
snyk-iac-test. This can be useful for snyk-iac-test to change its
behavioiur if this environment variable is set.